lobby/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.6.1
